Remove log file and clarifying last commit
This is a new page for assignment 4 that adds the ability to add and remove movies from the list. This list is saved in a text file and persisted. While saving to the text file is not the ideal way to do this, it is what the assignment asked for, so that's what I did.